Updated rchr_collectives.h

+++ rchr_collectives.h
@@ -23,6 +23,6 @@ void RecursiveInitCollectives(int device);
 // The ring allreduce. The lengths of the data chunks passed to this function
 // must be the same across all MPI processes. The output memory will be
 // allocated and written into `output`.
-void RecursiveAllreduce(float* data, size_t length, float** output);
+void RecursiveAllreduce(int8_t* data, size_t length, int8_t** output);
 
 #endif /* ifndef BAIDU_ALLREDUCE_RH_COLLECTIVES_H_ */
